- fixed more things.

+++ JoyceCode/engine/joyce/components/AnimationState.cs
@@ -5,6 +5,6 @@ public struct AnimationState
     public ModelAnimation? ModelAnimation;
     public ushort ModelAnimationFrame;
 
-    public const uint IsOneShot = 1;
+    public const ushort IsOneShot = 1;
     public ushort Flags;
 }
